Extracted security aspect of LogbookAutoConfiguration
Fixes #46
